backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlmodel>=0.0.8
feedparser>=6.0.10
apscheduler>=3.10.1
httpx>=0.24.1
beautifulsoup4>=4.12.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-magic>=0.4.27
Pillow>=10.0.0
aiofiles>=23.2.1
clamd>=1.0.2
PyYAML>=6.0.1
watchdog>=4.0.0
alembic>=1.13.0
